requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.1
python-dotenv==1.0.1
praw==7.7.1
httpx==0.26.0
pytest==8.0.0
